Move stuff around, allow variables to be null
